Vue 3 Lightbox
- Rename NewLightbox  MediaGallery/Lightbox
- Use previously committed refactor to typescript & composition api
- add story and spec
- replace vue-glide with vue3-carousel
